requirements.txt
streamlit

langchain-core
langchain-community
langchain-text-splitters

langchain-huggingface
sentence-transformers
torch

faiss-cpu

langchain-groq

pypdf

requests
beautifulsoup4

python-dotenv

numpy
tqdm
